backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
google-genai>=1.5.0
python-multipart==0.0.9
python-dotenv==1.0.0
pillow==10.2.0
pydantic==2.6.1